Intermediate commit

+++ src/UITypes.cpp
@@ -1,5 +1,5 @@
 #include "UITypes.h"
 
-UIPoint_t       defaultUIPoint      = {0,0};
+UIPoint_t       defaultUIPoint      = {(int16_t)0,(int16_t)0};
 UIDimensions_t  defaultUIDimensions = {defaultUIPoint,defaultUIPoint};
 UIColor_t       defaultUIColor      = {-1,-1,-1};
